+++ sdk/appconfiguration/app-configuration/CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ### Features Added
 
-- Add `audience` in `AppConfigurationClientOptions` and `KnownAppConfigurationAudience` to support specifying the token credential's Microsoft Entra audience when creating a client. When unspecified, the SDK will check the resource endpoint to determine the correct audience.
+- Add `audience` in `AppConfigurationClientOptions` and `KnownAppConfigurationAudience` to support specifying the token credential's Microsoft Entra audience when creating a client. When unspecified, the SDK will default to Azure Public Cloud.
 
 ### Breaking Changes